GS-946: Move report event helper into util class
- replace old helper with an autoloaded utility class
- update report pages and components to use it
- remove unnecessary guard checks
